Some clean up and RoutePlanner changes.

+++ src/dk/itu/n/danmarkskort/backend/OSMParser.java
@@ -40,7 +40,7 @@ public class OSMParser extends SAXAdapter {
     private transient boolean motorvehicle;
     private transient boolean walk;
     private transient boolean toGraph;
-    private transient int maxSpeed;
+    private transient short maxSpeed;
     private transient RouteController route;
     private transient HashMap<Point2D.Float, RouteVertex> vertexMap;
     private transient boolean waysFinished = false;
@@ -592,7 +592,7 @@ public class OSMParser extends SAXAdapter {
                 break;
             case "maxspeed":
                 try{
-                    maxSpeed = Integer.parseInt(v);
+                    maxSpeed = Short.parseShort(v);
                 }
                 catch (NumberFormatException e) {
                     maxSpeed = 0;
